Eine Warnungen ruhiggestellt

+++ quake2/src/server/sv_main.c
@@ -1030,7 +1030,7 @@ void SV_Shutdown (char *finalmsg, qboolean reconnect)
 
 	// free current level
 	if (sv.demofile)
-		FS_FCloseFile (sv.demofile);
+		FS_FCloseFile ((size_t)sv.demofile);
 	memset (&sv, 0, sizeof(sv));
 	Com_SetServerState (sv.state);
 
@@ -1040,7 +1040,7 @@ void SV_Shutdown (char *finalmsg, qboolean reconnect)
 	if (svs.client_entities)
 		Z_Free (svs.client_entities);
 	if (svs.demofile)
-		FS_FCloseFile (svs.demofile);
+		FS_FCloseFile ((size_t)svs.demofile);
 	memset (&svs, 0, sizeof(svs));
 }
